Use mmap to optimize circular buffer when parsing
Mmap two adjacent pages that point to the same physical memory. This
avoids length issue when benchmarking with ApacheBench, and also
increase rps of keep-alive request.

Somehow rps would drop when benchmarking with htstress, whose requests
aren't keep-alive. It is due to the inefficiency of mmap. Therefore a
more suitable solution should be considered.